refactor: fluent builder pattern for generated classes
also use long/Long instead of BigInteger

+++ src/main/java/com/sipgate/li/lib/x1/protocol/error/ErrorResponseException.java
@@ -1,7 +1,5 @@
 package com.sipgate.li.lib.x1.protocol.error;
 
-import java.math.BigInteger;
-
 /*
 
 ================================================================================
@@ -40,77 +38,77 @@ public abstract class ErrorResponseException extends Exception {
 
   // We've only implemented the exceptions that our implementation uses. Feel free to add more if required.
   // Check for the suggested details in ETSI TS 103 221-1, Table 46: Error codes
-  public static final BigInteger GENERIC_ERROR = BigInteger.valueOf(1000);
-  public static final BigInteger SYNTAX_SCHEMA_ERROR = BigInteger.valueOf(1010);
-  public static final BigInteger UNSUPPORTED_VERSION = BigInteger.valueOf(1020);
-  public static final BigInteger ADMF_IDENTIFIER_DOES_NOT_MATCH_CERTIFICATE_DETAILS = BigInteger.valueOf(1030);
-  public static final BigInteger UNEXPECTED_ADMF_IDENTIFIER = BigInteger.valueOf(1040);
-  public static final BigInteger NE_IDENTIFIER_DOES_NOT_MATCH_CERTIFICATE_DETAILS = BigInteger.valueOf(1050);
-  public static final BigInteger UNEXPECTED_NE_IDENTIFIER = BigInteger.valueOf(1060);
-  public static final BigInteger KEEPALIVE_NOT_SUPPORTED = BigInteger.valueOf(1070);
-  public static final BigInteger UNSUPPORTED_REQUEST = BigInteger.valueOf(1080);
+  public static final long GENERIC_ERROR = 1000L;
+  public static final long SYNTAX_SCHEMA_ERROR = 1010L;
+  public static final long UNSUPPORTED_VERSION = 1020L;
+  public static final long ADMF_IDENTIFIER_DOES_NOT_MATCH_CERTIFICATE_DETAILS = 1030L;
+  public static final long UNEXPECTED_ADMF_IDENTIFIER = 1040L;
+  public static final long NE_IDENTIFIER_DOES_NOT_MATCH_CERTIFICATE_DETAILS = 1050L;
+  public static final long UNEXPECTED_NE_IDENTIFIER = 1060L;
+  public static final long KEEPALIVE_NOT_SUPPORTED = 1070L;
+  public static final long UNSUPPORTED_REQUEST = 1080L;
 
-  public static final BigInteger XID_ALREADY_EXISTS = BigInteger.valueOf(2010);
-  public static final BigInteger XID_DOES_NOT_EXIST = BigInteger.valueOf(2020);
-  public static final BigInteger DID_ALREADY_EXISTS = BigInteger.valueOf(2030);
-  public static final BigInteger DID_DOES_NOT_EXIST = BigInteger.valueOf(2040);
-  public static final BigInteger GENERICOBJECTID_ALREADY_EXISTS = BigInteger.valueOf(2050);
-  public static final BigInteger GENERICOBJECTID_DOES_NOT_EXIST = BigInteger.valueOf(2060);
+  public static final long XID_ALREADY_EXISTS = 2010L;
+  public static final long XID_DOES_NOT_EXIST = 2020L;
+  public static final long DID_ALREADY_EXISTS = 2030L;
+  public static final long DID_DOES_NOT_EXIST = 2040L;
+  public static final long GENERICOBJECTID_ALREADY_EXISTS = 2050L;
+  public static final long GENERICOBJECTID_DOES_NOT_EXIST = 2060L;
 
-  public static final BigInteger GENERIC_ACTIVATETASK_FAILURE = BigInteger.valueOf(3000);
-  public static final BigInteger GENERIC_MODIFYTASK_FAILURE = BigInteger.valueOf(3001);
-  public static final BigInteger UNSUPPORTED_TARGETIDENTIFIER_TYPE = BigInteger.valueOf(3010);
-  public static final BigInteger UNSUPPORTED_COMBINATION_OF_TARGETIDENTIFIERS = BigInteger.valueOf(3020);
-  public static final BigInteger MULTIPLE_DESTINATIONS_NOT_SUPPORTED = BigInteger.valueOf(3030);
-  public static final BigInteger INVALID_COMBINATION_OF_DELIVERYTYPE_AND_DESTINATIONS = BigInteger.valueOf(3040);
+  public static final long GENERIC_ACTIVATETASK_FAILURE = 3000L;
+  public static final long GENERIC_MODIFYTASK_FAILURE = 3001L;
+  public static final long UNSUPPORTED_TARGETIDENTIFIER_TYPE = 3010L;
+  public static final long UNSUPPORTED_COMBINATION_OF_TARGETIDENTIFIERS = 3020L;
+  public static final long MULTIPLE_DESTINATIONS_NOT_SUPPORTED = 3030L;
+  public static final long INVALID_COMBINATION_OF_DELIVERYTYPE_AND_DESTINATIONS = 3040L;
 
-  public static final BigInteger UNSUPPORTED_SERVICETYPE = BigInteger.valueOf(3050);
+  public static final long UNSUPPORTED_SERVICETYPE = 3050L;
 
-  public static final BigInteger GENERIC_DEACTIVATETASK_FAILURE = BigInteger.valueOf(4000);
+  public static final long GENERIC_DEACTIVATETASK_FAILURE = 4000L;
 
-  public static final BigInteger GENERIC_DEACTIVEALLTASKS_FAILURE = BigInteger.valueOf(5000);
-  public static final BigInteger DEACTIVATEALLTASKS_NOT_ENABLED = BigInteger.valueOf(5010);
+  public static final long GENERIC_DEACTIVEALLTASKS_FAILURE = 5000L;
+  public static final long DEACTIVATEALLTASKS_NOT_ENABLED = 5010L;
 
-  public static final BigInteger GENERIC_CREATEDESTINATION_FAILURE = BigInteger.valueOf(6000);
-  public static final BigInteger GENERIC_MODIFYDESTINATION_FAILURE = BigInteger.valueOf(6001);
-  public static final BigInteger UNSUPPORTED_DELIVERYADDRESS_TYPE = BigInteger.valueOf(6020);
+  public static final long GENERIC_CREATEDESTINATION_FAILURE = 6000L;
+  public static final long GENERIC_MODIFYDESTINATION_FAILURE = 6001L;
+  public static final long UNSUPPORTED_DELIVERYADDRESS_TYPE = 6020L;
 
-  public static final BigInteger GENERIC_REMOVEDESTINATION_FAILURE = BigInteger.valueOf(7000);
-  public static final BigInteger DESTINATION_IN_USE = BigInteger.valueOf(7010);
+  public static final long GENERIC_REMOVEDESTINATION_FAILURE = 7000L;
+  public static final long DESTINATION_IN_USE = 7010L;
 
-  public static final BigInteger GENERIC_REMOVEALLDESTINATIONS_FAILURE = BigInteger.valueOf(8000);
-  public static final BigInteger DESTINATIONS_IN_USE = BigInteger.valueOf(8010);
-  public static final BigInteger REMOVEALLDESTINATIONS_NOT_ENABLED = BigInteger.valueOf(8020);
+  public static final long GENERIC_REMOVEALLDESTINATIONS_FAILURE = 8000L;
+  public static final long DESTINATIONS_IN_USE = 8010L;
+  public static final long REMOVEALLDESTINATIONS_NOT_ENABLED = 8020L;
 
-  public static final BigInteger GENERIC_CREATEOBJECT_FAILURE = BigInteger.valueOf(8500);
-  public static final BigInteger GENERIC_MODIFYOBJECT_FAILURE = BigInteger.valueOf(8510);
-  public static final BigInteger GENERIC_DELETEOBJECT_FAILURE = BigInteger.valueOf(8530);
-  public static final BigInteger GENERIC_DELETEALLOBJECTS_FAILURE = BigInteger.valueOf(8540);
-  public static final BigInteger DELETEALLOBJECTS_NOT_ENABLED = BigInteger.valueOf(8550);
+  public static final long GENERIC_CREATEOBJECT_FAILURE = 8500L;
+  public static final long GENERIC_MODIFYOBJECT_FAILURE = 8510L;
+  public static final long GENERIC_DELETEOBJECT_FAILURE = 8530L;
+  public static final long GENERIC_DELETEALLOBJECTS_FAILURE = 8540L;
+  public static final long DELETEALLOBJECTS_NOT_ENABLED = 8550L;
 
-  public static final BigInteger ERROR_CLEARED = BigInteger.valueOf(9000);
-  public static final BigInteger GENERIC_WARNING = BigInteger.valueOf(9010);
-  public static final BigInteger GENERIC_NON_TERMINATING_FAULT = BigInteger.valueOf(9020);
-  public static final BigInteger TERMINATING_FAULT = BigInteger.valueOf(9030);
-  public static final BigInteger REQUEST_ACTIONED = BigInteger.valueOf(9040);
-  public static final BigInteger KEEPALIVES_NOT_RECEIVED = BigInteger.valueOf(9050);
+  public static final long ERROR_CLEARED = 9000L;
+  public static final long GENERIC_WARNING = 9010L;
+  public static final long GENERIC_NON_TERMINATING_FAULT = 9020L;
+  public static final long TERMINATING_FAULT = 9030L;
+  public static final long REQUEST_ACTIONED = 9040L;
+  public static final long KEEPALIVES_NOT_RECEIVED = 9050L;
 
-  public static final BigInteger DATABASE_CLEARED = BigInteger.valueOf(10000);
+  public static final long DATABASE_CLEARED = 10000L;
 
-  private final BigInteger errorCode;
+  private final long errorCode;
   private final String errorDescription;
 
-  protected ErrorResponseException(final BigInteger errorCode) {
+  protected ErrorResponseException(final long errorCode) {
     this(errorCode, null);
   }
 
-  protected ErrorResponseException(final BigInteger errorCode, final String errorDescription) {
+  protected ErrorResponseException(final long errorCode, final String errorDescription) {
     super(errorDescription);
     this.errorCode = errorCode;
     this.errorDescription = errorDescription;
   }
 
-  public BigInteger getErrorCode() {
+  public long getErrorCode() {
     return errorCode;
   }
