+++ src/main/java/com/sipgate/li/lib/x1/server/entity/Destination.java
@@ -8,5 +8,5 @@ public record Destination(
   String friendlyName,
   DeliveryType deliveryType,
   String deliveryAddress,
-  int deliveryPort
+  long deliveryPort
 ) {}
